PB-30 digraph class is added (#44)
* digraph class is added

* deleted comment

+++ src/main/java/co/edu/uptc/model/PropertiesEdge.java
@@ -64,7 +64,7 @@ public class PropertiesEdge implements Documentable {
     this.junction = junction;
   }
 
-  public double getStartNode() {
+  public int getStartNode() {
     return startNode;
   }
